refactor: rename P2 violations in route-plugin and location-plugin (#1215)
route-plugin (5 files):
- draft→draftUtils, osrm-defaults→osrmDefaultUtils, ideGsmCsv→ideGsmCsvUtils,
  filteringSamples→filteringSampleConstants, progress→progressTypes

location-plugin (12 files):
- optional-features→optionalFeatureUtils, plugin-worker-factory→pluginWorkerFactoryUtils,
  runtime-worker→runtimeWorkerUtils, payloads→payloadTypes, csvSources→csvSourceConstants,
  csvUtils→csvDownloadUtils, mappers→mapperUtils, strategyRegistry→strategyRegistryUtils,
  pointFactories→pointFactoryUtils, ideGsmProgress→ideGsmProgressUtils,
  ideGsmSelection→ideGsmSelectionUtils, normalizers→normalizerUtils

Rollback: git revert <hash> --no-edit
