+++ plugins/location-plugin/src/common/test-shims/optionalFeatureUtils.ts
