+++ plugins/location-plugin/src/common/test-shims/pluginWorkerFactoryUtils.ts
