+++ plugins/location-plugin/src/common/types/payloadTypes.ts
