+++ plugins/location-plugin/src/services/download/__tests__/unit/csvSources.unit.test.ts
@@ -4,7 +4,7 @@ import {
   parseOpenFlightsCsv,
   parseOurAirportsCsv,
   parseWorldPortIndexCsv,
-} from '../../csvSources';
+} from '../../csvSourceConstants';
 
 describe('csvSources', () => {
   const timestamp = Date.now() as Timestamp;
