+++ plugins/location-plugin/src/services/download/mapperUtils.ts
