+++ plugins/location-plugin/src/services/download/strategyRegistryUtils.ts
