+++ plugins/location-plugin/src/ui/state/ideGsmProgressUtils.ts
