+++ plugins/location-plugin/src/ui/utils/ideGsmSelectionUtils.ts
