+++ plugins/location-plugin/src/worker/createLocationFeatureStoreDexie.ts
@@ -2,7 +2,7 @@ import type { NodeId } from '@hierarchidb/core-types';
 import type { FeatureItemBase, FeatureStore } from '@hierarchidb/runtime-worker';
 import type { LocationDB } from './locationEntitiesDB.js';
 import type { LocationGroupItemData } from '~/common/types/entities';
-import { fromGroupRow, toGroupRow } from './normalizers.js';
+import { fromGroupRow, toGroupRow } from './normalizerUtils.js';
 
 type Item = FeatureItemBase<LocationGroupItemData>;
