+++ plugins/route-plugin/src/common/utils/draftUtils.ts
